+++ NEWS.md
@@ -1,3 +1,16 @@
+# sigminer 1.0.5
+
+- Exported APOBEC result when the mode is 'ALL' in `sig_tally()`.
+- Added batch bootstrap analysis feature (#158).
+- Supported all common signature plotting.
+- Added strand feature to signature profile.
+
+# sigminer 1.0.4
+
+- Added profile plot for DBS and INDEL.
+- Fixed error for signature extraction in mode 'DBS' or 'ID'.
+- Fixed method 'M' for CN tally cannot work when `cores > 1` (#161).
+
 # sigminer 1.0.3
 
 - Added multiple methods for `sig_fit()`.